+++ lib/optimizer.ml
@@ -6,6 +6,8 @@ let rec optimize (e: expr) : expr = match e with
   | Mult(Integer x, Integer y) -> Integer (x * y)
   | Eq(Boolean x, Boolean y) -> Boolean (x == y)
   | Eq(x, y) -> Eq (optimize x, optimize y)
+  | Gt(x, y) -> Gt (optimize x, optimize y)
+  | Lt(x, y) -> Lt (optimize x, optimize y)
   | And(Boolean x, Boolean y) -> Boolean (x && y) (* Optimize bool primitives *)
   | Or(Boolean x, Boolean y) -> Boolean (x || y)
   | Not(Boolean x) -> Boolean (not x)
@@ -24,11 +26,13 @@ let rec optimize (e: expr) : expr = match e with
   | Lambda(params, body) -> Lambda(params, optimize body)
   | Let(declarations, body) -> optimize_let declarations body false
   | Letlazy(declarations, body) -> optimize_let declarations body true
+  | Letrec(ident, dec, body) -> Letrec(ident, optimize dec, optimize body)
+  | Letreclazy(ident, dec, body) -> Letreclazy(ident, optimize dec, optimize body)
   | Apply(func, params) -> Apply(func, (List.map optimize params))
   | _ -> e
   and optimize_let declarations body islazy =
   let od = List.map (fun (i, v) -> (i, optimize v)) declarations in
-    let ob = optimize body in
+  let ob = optimize body in
     if List.length od = 1 then
       let (ident, value) = List.hd od in
       if ob = Symbol ident